test(api): reset jobs helper and update imports
